+++ game-server/data/scripts/system/handlers/ai/instance/tiamatStrongHold/MuraganAI2.java
@@ -95,7 +95,7 @@ public class MuraganAI2 extends GeneralNpcAI2 {
   	WorldMapInstance instance = getOwner().getPosition().getWorldMapInstance();
   	for (Npc npc : instance.getNpcs()) {
   		if (npc.getNpcId() == 219392) {//4.0
-  			spawn(283145, npc.getX(), npc.getY(), npc.getZ(), (byte) npc.getHeading());//4.0
+  			spawn(283145, npc.getX(), npc.getY(), npc.getZ(), npc.getHeading());//4.0
   			npc.getController().onDelete();
   		}
   	}
