+++ game-server/src/com/aionemu/gameserver/model/utils3d/Point3D.java
@@ -22,9 +22,9 @@ public class Point3D {
 	}
 
 	public Point3D(float x, float y, float z) {
-		this.x = (double) x;
-		this.y = (double) y;
-		this.z = (double) z;
+		this.x = x;
+		this.y = y;
+		this.z = z;
 	}
 
 	public double distance(Point3D p) {
